Move RunResources struct to vm::runners::cairo_runner::RunResources
